:recycle: [Refactor] openUserModal  함수 분리 #main
